Refactor. SOLID try. Timings added.
